Using pre-calculated hash instead of entire header.
Letting the host compute the first part of the hash (the
version, previous block hash, and first 7 words of merkle tree).
This does not take much compute time, and saves quite many registers.
(32x16 in wishbone module, 32x16 in syncer, and 32x8 for storing the
first part of the hash).